+++ cmd/datawatch/main.go
@@ -79,7 +79,7 @@ import (
 )
 
 // Version is set at build time via -ldflags.
-var Version = "3.2.0"
+var Version = "3.3.0"
 
 var (
 	cfgPath    string